Rename 'create-config.js' to 'karma-config.js' for clarity
